make v8 behind a build flag
