+++ db/database_no_v8.go
@@ -1,0 +1,6 @@
+// -build cb_sg_v8
+
+package db
+
+// The default JavaScript engine if not otherwise specified in the DatabaseContextOptions.
+const DefaultJavaScriptEngine = "Otto" // Can be "Otto" or "V8"
