+++ db/functions/graphql_test.go
@@ -1,3 +1,6 @@
+//go:build cb_sg_v8
+// +build cb_sg_v8
+
 /*
 Copyright 2020-Present Couchbase, Inc.
