+++ js/service_no_v8.go
@@ -1,0 +1,9 @@
+// -build cb_sg_v8
+
+package js
+
+type V8BasicTemplate struct {
+}
+
+type V8Template struct {
+}
