+++ js/test_utils_no_v8.go
@@ -1,0 +1,4 @@
+package js
+
+// -build cb_sg_v8
+var testEngines = []*Engine{Otto}
