+++ js/test_utils_v8.go
@@ -1,0 +1,6 @@
+//go:build cb_sg_v8
+// +build cb_sg_v8
+
+package js
+
+var testEngines = []*Engine{V8, Otto}
