+++ js/vmpool_test.go
@@ -1,3 +1,6 @@
+//go:build cb_sg_v8
+// +build cb_sg_v8
+
 /*
 Copyright 2022-Present Couchbase, Inc.
