+++ rest/functionsapitest/user_functions_queries_test.go
@@ -6,6 +6,9 @@
 //  software will be governed by the Apache License, Version 2.0, included in
 //  the file licenses/APL2.txt.
 
+//go:build cb_sg_v8
+// +build cb_sg_v8
+
 package functionsapitest
 
 import (
